port check 017: check_main_entries_in_the_name_table
(issue #1388)
