+++ bin/fontbakery-check-ttf.py
@@ -197,7 +197,6 @@ def fontbakery_check_ttf(config):
     checks.check_OS2_xAvgCharWidth(fb, font)
 
 #    checks.check_copyright_entries_match_license(fb, found, file_path, font)
-    checks.check_main_entries_in_the_name_table(fb, font, target)
     checks.check_name_entries_symbol_substitutions(fb, font)
     checks.check_description_strings_do_not_exceed_100_chars(fb, font)
